# docs/swagger.yaml
definitions:
  entity.Payment:
    type: object
  response.APIResponse:
    properties:
      data:
        description: actual payload
      entity:
        description: e.g. "payments"
        type: string
      message:
        description: e.g. "Success Get All Payments"
        type: string
      state:
        description: e.g. "getAllPayments"
        type: string
      status:
        description: '"success" or "failed"'
        type: string
    type: object
host: localhost:8080
info:
  contact:
    name: ADF Code
    url: https://github.com/adf-code
  description: API service to manage payment using Clean Architecture
  title: Beta Payment API
  version: "1.0"
paths:
  /api/v1/payments:
    get:
      consumes:
      - application/json
      description: List all payments with filter, search, pagination
      parameters:
      - description: Search field (e.g., title)
        in: query
        name: search_field
        type: string
      - description: Search value (e.g., golang)
        in: query
        name: search_value
        type: string
      - collectionFormat: multi
        description: Filter field
        in: query
        items:
          type: string
        name: filter_field
        type: array
      - collectionFormat: multi
        description: Filter value
        in: query
        items:
          type: string
        name: filter_value
        type: array
      - collectionFormat: multi
        description: Range field
        in: query
        items:
          type: string
        name: range_field
        type: array
      - collectionFormat: multi
        description: Range lower bound
        in: query
        items:
          type: string
        name: from
        type: array
      - collectionFormat: multi
        description: Range upper bound
        in: query
        items:
          type: string
        name: to
        type: array
      - description: Sort field
        in: query
        name: sort_field
        type: string
      - description: Sort direction ASC/DESC
        in: query
        name: sort_direction
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Limit per page
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Get list of payments
      tags:
      - payments
    post:
      consumes:
      - application/json
      description: Creates a new payment with the status auto generate to pending
      parameters:
      - description: Payment data to create
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/entity.Payment'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.APIResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Create a new payment
      tags:
      - payments
  /api/v1/payments/{id}:
    delete:
      description: Deletes a payment entity using its UUID
      parameters:
      - description: UUID of the payment to delete
        in: path
        name: id
        required: true
        type: string
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/response.APIResponse'
        "400":
          description: Invalid UUID
          schema:
            $ref: '#/definitions/response.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.APIResponse'
        "404":
          description: Payment not found
          schema:
            $ref: '#/definitions/response.APIResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Delete a payment by ID
      tags:
      - payments
    get:
      description: Modify a payment entity using its UUID
      parameters:
      - description: UUID of the payment
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.APIResponse'
        "400":
          description: Invalid UUID
          schema:
            $ref: '#/definitions/response.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.APIResponse'
        "404":
          description: Payment not found
          schema:
            $ref: '#/definitions/response.APIResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Update payment by ID
      tags:
      - payments
  /healthz:
    get:
      description: Health check for service
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.APIResponse'
      summary: Health Check
      tags:
      - health
securityDefinitions:
  BearerAuth:
    description: 'Using token header using the Bearer scheme. Example: "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"